+++ examples/src/example.js
@@ -24,15 +24,17 @@ import {InsertLinkModal} from './insertLink';
 import {EmojiView} from './emoji';
 
 const initHTML = `<br/>
-<center><b ontouchstart="_.sendEvent('TitleClick')" id="title">Rich Editor</b></center>
+<center><b onclick="_.sendEvent('TitleClick')" id="title" >Rich Editor</b></center>
 <center>
 <a href="https://github.com/wxik/react-native-rich-editor">React Native</a>
 <span>And</span>
 <a href="https://github.com/wxik/flutter-rich-editor">Flutter</a>
 </center>
 <br/>
-<img src="https://upload.wikimedia.org/wikipedia/commons/thumb/a/a7/React-icon.svg/120px-React-icon.svg.png" ontouchstart="_.sendEvent('ImgClick')"/>
-<br/><br/><br/><br/>
+<img src="https://upload.wikimedia.org/wikipedia/commons/thumb/a/a7/React-icon.svg/120px-React-icon.svg.png"  onclick="_.sendEvent('ImgClick')"/>
+<br/><br/>
+<span ontouchstart="event.preventDefault();" ontouchend="_.sendEvent('SwitchImage')"> Click Switch Image</span>
+<br/><br/>
 `;
 
 const phizIcon = require('./assets/phiz.png');
@@ -134,6 +136,7 @@ class Example extends React.Component {
     insertVideo() {
         this.richText.current?.insertVideo(
             'https://mdn.github.io/learning-area/html/multimedia-and-embedding/video-and-audio-content/rabbit320.mp4',
+            'width: 50%;',
         );
     }
 
@@ -145,6 +148,7 @@ class Example extends React.Component {
         // insert URL
         this.richText.current?.insertImage(
             'https://upload.wikimedia.org/wikipedia/commons/thumb/a/a7/React-icon.svg/100px-React-icon.svg.png',
+            'background: gray;',
         );
         // insert base64
         // this.richText.current?.insertImage(`data:${image.mime};base64,${image.data}`);
@@ -204,22 +208,30 @@ class Example extends React.Component {
         console.log('KeyDown:', data);
     };
 
-    handleMessage = ({type, data}) => {
+    handleMessage = ({type, id, data}) => {
         switch (type) {
             case 'ImgClick':
+                break;
+            case 'TitleClick':
                 const index = this._tempIndex || 0;
                 const color = ['red', 'blue', 'gray', 'yellow', 'coral'][index];
                 this._tempIndex = index + 1 >= color.length ? 0 : index + 1;
 
                 // command: $ = document.querySelector
-                this.richText.current?.commandDOM(`$('#title').style.color='${color}'`);
+                this.richText.current?.commandDOM(`$('#${id}').style.color='${color}'`);
                 break;
-            case 'TitleClick':
-                Alert.alert('Click Title');
+            case 'SwitchImage':
                 break;
-            default:
-                console.log('onMessage', type, data);
         }
+        console.log('onMessage', type, id, data);
+    };
+
+    handleFocus = () => {
+        console.log('Editor Focus');
+    };
+
+    handleBlur = () => {
+        console.log('Editor Blur');
     };
 
     render() {
@@ -280,6 +292,8 @@ class Example extends React.Component {
                         onKeyUp={that.handleKeyUp}
                         onKeyDown={that.handleKeyDown}
                         onMessage={that.handleMessage}
+                        onFocus={that.handleFocus}
+                        onBlur={that.handleBlur}
                         pasteAsPlainText={true}
                     />
                 </ScrollView>
@@ -343,6 +357,8 @@ const styles = StyleSheet.create({
     rich: {
         minHeight: 300,
         flex: 1,
+        borderBottomWidth: StyleSheet.hairlineWidth,
+        borderColor: '#e8e8e8',
     },
     richBar: {
         height: 50,